feat: puluse sesesion card skeleton
